requirements.txt
python-telegram-bot==13.1
emoji==0.6.0
